Extract shared integration test helpers into utils/integrationTestHelpers.ts
Move withTempDir, isExecutableAvailable, and isRenvAvailable from
publish/integration.test.ts and inspect/integration.test.ts into a
shared module to reduce duplication across integration test files.
